+++ mambular/models/fttransformer_classifier.py
@@ -1,9 +1,9 @@
-from .sklearn_base_regressor import SklearnBaseRegressor
+from .sklearn_base_classifier import SklearnBaseClassifier
 from ..base_models.ft_transformer import FTTransformer
-from ..utils.configs import DefaultFTTransformerConfig
+from ..configs.fttransformer_config import DefaultFTTransformerConfig
 
 
-class FTTransformerRegressor(SklearnBaseRegressor):
+class FTTransformerClassifier(SklearnBaseClassifier):
     def __init__(self, **kwargs):
         super().__init__(
             model=FTTransformer, config=DefaultFTTransformerConfig, **kwargs
